pyproject.toml
[tool.poetry]
name = "3d_tennis_lab"
version = "0.0.1"
description = "A project for 3D analysis of tennis matches, tracking court, players, and ball."
authors = ["motoki <kamimura098@gamil.com>"]
repository = "https://github.com/motoki/3d-tennis-lab"
documentation = "https://motoki.github.io/3d-tennis-lab/"
readme = "README.md"
packages = [{ include = "3d_tennis_system" }]

[[tool.poetry.source]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[tool.poetry.dependencies]
# 3点セットを cu121 ソースから取得
torch = { version = "2.3.0", source = "pytorch-cu121" }
torchvision = { version = "0.22.0", source = "pytorch-cu121" }
torchaudio = { version = "2.3.0", source = "pytorch-cu121" }

pandas = "^2.3.1"
scikit-learn = "^1.7.1"
matplotlib = "^3.10.5"
seaborn = "^0.13.2"
pillow = "^11.3.0"
tensorboard = "^2.20.0"
tqdm = "^4.67.1"
hydra-core = "^1.3.2"
pytorch-lightning = "^2.5.2"
opencv-python = "^4.12.0.88"
numpy = "^2.2.6"
jupyterlab = "^4.2.3"
albumentations = "^2.0.8"
timm = "^1.0.19"
transformers = "^4.55.0"
dotmap = "^1.3.30"
rich = "13.7.1"
yt-dlp = "^2025.8.11"

# --- 4D-Humans 関連（ビルド事故を避けるため「別グループ」に退避） ---
# ※ chumpy は NumPy 2.x と相性が悪く PEP517 ビルドで落ちやすいので、
#   既定インストールからは外し、使うときだけ group オプションで入れる。
#   （detectron2 も CUDA 構成依存が強いため同様）
# 使うとき: `poetry install --with humans4d`


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"
tox = "^4.15.1"
ipykernel = "^6.29.4"
matplotlib = "^3.9.0"
Pillow = "^10.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["3d_tennis_system"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.pytest.ini_options]
testpaths = ["tests", "development", "experiments"]
addopts = "--import-mode=importlib --cov=3d_tennis_system --cov=development --cov=experiments --cov-branch --cov-report=xml"
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true
exclude = [".venv", "build", "dist", ".mypy_cache", ".pytest_cache", "site", "docs/_site"]

[tool.ruff.format]
preview = true

[tool.ruff.lint]
select = ["E","W","F","I","UP","B","C4","SIM","RUF"]
ignore = ["E501", "F841"]
preview = true

[tool.ruff.lint.per-file-ignores]
"tools/**" = ["E","W","I","UP","B","C4","SIM","RUF"]
"experiments/**" = ["E","W","I","UP","B","C4","SIM","RUF"]
"development/**" = ["E","W","I","UP","B","C4","SIM","RUF"]
"developments/**" = ["E","W","I","UP","B","C4","SIM","RUF"]
"**/tests/**" = ["E","W","I","UP","B","C4","SIM","RUF"]

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["3d_tennis_system", "development", "experiments"]
omit = [
  "config.py",
  "config-*.py",
  "*/.venv/*",
]
